ecommerce/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-dotenv==1.0.1
